cloud_functions/functions/requirements.txt
firebase-functions
cloudevents
Pillow